Pothole-Backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
ultralytics>=8.0
Pillow>=10.0
python-dotenv>=1.0
opencv-python-headless>=4.8
numpy>=1.24
requests>=2.31
django-cors-headers>=4.3